pyproject.toml
[tool.poetry]
name = "telegram-mp3"
version = "0.1.0"
description = ""
authors = ["Koryws <sidorenkodk@gmail.com>"]
readme = "README.md"
packages = [{include = "telegram_mp3"}]

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^2.22.1"
gTTS = "^2.2.4"
pdfplumber = "^0.7.4"
python-dotenv = "^0.21.0"
python-docx = "^0.8.11"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
